Move UserAPIKeys to components/admin
